Moving some code from LongShort contract to yield manager contract
Variable moved: totalValueLockedInYieldManager
There was only 1 place this variable was being used, and it was being used to
keep track of how much yield is reserved for the Treasury. I reckon this
calculation & storage is better suited to living in the yield managers.

Function moved: claimAndDistributeYield
It seems more fitting to have this code live in the yield manager.

======================

Before this change: the LongShort contract remembered how much yield is owed to
the Treasury, and it also handled the code that updated this value.

After this change: the LongShort contract remembers how much fees are owed to
the Treasury, and the YieldManager contract remembers how much yield is owed to
the Treasury, as well as code for updating this value.

Also, this commit splits out the getTreasurySplit/getMarketSplit functions so
that there are now functions that will give the percentage split.

Also, removed treausuryOnly modifier from transferTreasuryFunds function. It
doesn't matter who calls this function, the funds will go to the treasury.

Changed the names of some functions/variables.