Removing syntheticTokenPrice mapping variables from LongShort
_refreshTokenPrice is a source of errors because one has to be careful to call
it EITHER everytime one it's inputs changes (syntheticTokenBackedValue,
longTokenSupply, shortTokenSupply) OR when it's output
changes (syntheticTokenPrice). It's not entirely obvious that this needs to be
done so I tried to make this function safer. While investigating I realized that
there isn't actually a need to store the synthetic token prices. They can just
be calculated on the fly.

The trade-off here is that the calculation needs to be done twice now in
_updateSystemStateInternal instead of just once, as was the case before this
change. So either we can calculate twice or we can store the result on chain.
Since the calculation is simple (1 multiply & 1 divide) it is cheaper gas-wise
to do this calculation twice than to store the result on-chain.

A nice side-effect of this change is that the code is simplified: the
_refreshTokenPrice function is not needed anymore and there is one less mapping
state variable.